nfl_scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pandas>=2.0.0
pyarrow>=12.0.0
pyyaml>=6.0
selenium>=4.8.0
webdriver-manager>=3.8.0
brotli>=1.0.0
playwright>=1.40.0

